requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
python-docx==1.1.0
plotly==5.18.0
pandas>=2.2,<3.0
gunicorn==21.2.0
werkzeug==3.0.1
